Don't load packages linearly
Changes the behaviour of package loading to be asynchronous so they show
up as soon as each package loads.

This seems to be better because normally there is a black screen for a
few seconds, before all the packages start loading, so at first glance
the page looks boring. Whereas now, packages will start appearing as
soon as they are ready.